major version update

+++ chiak-app/src/views/ShopList.vue
@@ -1,7 +1,7 @@
 <template>
-    <div>
+    <div class="bg-homepage bg-cover bg-fixed min-h-screen h-full pb-10">
         <Nav />
-        <div class="m-20 bg-lime-50 mx-36 pb-3 px-2">
+        <div class="mt-20 bg-lime-50 mx-36 pb-3 px-6 rounded-md">
             <div class="container flex flex-wrap justify-between items-center mx-auto">
                 <div class="flex flex-wrap ml-10 mt-10">
                     <img
@@ -17,7 +17,7 @@
             <!-- <div>
                 <button class="flex mr-20 mt-10 p-2 px-4 bg-lime-600 rounded-full text-white" @click="test()">Test</button>
             </div> -->
-            <Items @clear-product="deleteProduct" :items="items" />
+            <Items @clear-product="deleteProduct" :items="items" :display="display"/>
         </div>
     </div>
 </template> 
@@ -32,7 +32,7 @@ export default {
     name: 'ShopList',
     data() {
         return {
-            items: []
+            items: [],
         }
     },
     created() {
@@ -46,6 +46,7 @@ export default {
         })
     },
     mounted() {
+        
         if(!UserService.isLoggedIn()){
             this.$router.push('/login')
         }
@@ -60,20 +61,21 @@ export default {
             console.log(this.items.length)
         },
         clearAll() {
-            this.items = []
-            ShoppingListService.deleteAllProducts();
+            if (confirm("You are about to clear all items, are you sure?")){
+                ShoppingListService.deleteAllProducts()
+                .then(() => {
+                    this.items = []
+                })
+            }
         },
         deleteProduct(id) {
             // console.log('items', id);
-            this.items = this.items.filter((item) => item.id !== id)
-            ShoppingListService.deleteProduct(id);
-            // .then((res) => {
-            //     if(res == "failed") {
-
-            //     } else {
-            //         this.$router.push('/')
-            //     }
-            // });
+            if (confirm("You are about to clear this item, are you sure?")){
+                ShoppingListService.deleteProduct(id)
+                .then (() => {
+                    this.items = this.items.filter((item) => item.id !== id)
+                })
+            }
         },
     },
